+++ guardpost/authorization.py
@@ -7,14 +7,19 @@ from rodi import ContainerProtocol
 
 from guardpost.abc import BaseStrategy
 from guardpost.authentication import Identity
+from guardpost.errors import AuthException
 
 
-class AuthorizationError(Exception):
-    pass
+class AuthorizationError(AuthException):
+    """
+    Base class for all kinds of AuthorizationErrors.
+    """
 
 
-class AuthorizationConfigurationError(Exception):
-    pass
+class AuthorizationConfigurationError(AuthException):
+    """
+    Exception to describe errors in user-defined authorization configuration.
+    """
 
 
 class PolicyNotFoundError(AuthorizationConfigurationError, RuntimeError):